Change int to float in tests

+++ tests/test_file_schemas.py
@@ -5394,7 +5394,7 @@ class BlendAscwdsPirData:
             StructField(
                 IndCQC.people_directly_employed_filled_posts, FloatType(), True
             ),
-            StructField(IndCQC.ascwds_filled_posts_dedup_clean, IntegerType(), True),
+            StructField(IndCQC.ascwds_filled_posts_dedup_clean, FloatType(), True),
         ]
     )
     expected_blend_modelled_pir_ands_ascwds_schema = StructType(
@@ -5409,7 +5409,7 @@ class BlendAscwdsPirData:
             StructField(
                 IndCQC.people_directly_employed_filled_posts, FloatType(), True
             ),
-            StructField(IndCQC.ascwds_filled_posts_dedup_clean, IntegerType(), True),
+            StructField(IndCQC.ascwds_filled_posts_dedup_clean, FloatType(), True),
             StructField(IndCQC.ascwds_pir_merged, FloatType(), True),
         ]
     )
